Bug [2831201] - Robocode enters infinite loop with the Restart button. When the GUI is shown, the RobotThreadManager.disposeAppContext() sometimes took a long time to perform, causing the battle shutdown to hang for several seconds. This has been fixed, and thus restarting the battle with the GUI on is now faster.

+++ robocode.battle/src/main/java/net/sf/robocode/battle/Battle.java
@@ -146,7 +146,7 @@ public final class Battle extends BaseBattle {
 	// Turn skip related items
 	private boolean parallelOn;
 	static final int DEBUG_TURN_WAIT = 10 * 60 * 1000;
-	private int millisWait;
+	private long millisWait;
 	private int microWait;
 
 	// Objects in the battle
@@ -382,10 +382,10 @@ public final class Battle extends BaseBattle {
 			}
 			final long waitTime = (long) (cpuConstant * parallelConstant);
 
-			millisWait = (int) (waitTime / 1000000);
+			millisWait = waitTime / 1000000;
 			microWait = (int) (waitTime % 1000000);
 		} else {
-			millisWait = (int) (cpuConstant / 1000000);
+			millisWait = cpuConstant / 1000000;
 			microWait = (int) (cpuConstant % 1000000);
 		}
 		if (microWait == 0) {
