+++ robocode.battle/src/main/java/net/sf/robocode/battle/peer/IRobotPeerBattle.java
@@ -53,7 +53,7 @@ public interface IRobotPeerBattle extends ContestantPeer {
 
 	void waitWakeup();
 
-	void waitSleeping(int millisWait, int microWait);
+	void waitSleeping(long millisWait, int microWait);
 
 	void waitForStop();
